requirements.txt
streamlit~=1.40.0
openai
python-dotenv
numpy~=2.1.3
pillow~=11.0.0
GitPython~=3.1.43
gitdb~=4.0.11
pip~=23.2.1
attrs~=24.2.0
protobuf~=5.28.3
pytz~=2024.2
rich~=13.9.4
Pygments~=2.18.0
markdown-it-py~=3.0.0
toml~=0.10.2
typing_extensions~=4.12.2
smmap~=5.0.1
mdurl~=0.1.2
charset-normalizer~=3.4.0
altair~=5.4.1
jsonschema~=4.23.0
narwhals~=1.13.3
pandas~=2.2.3
pyarrow~=18.0.0
Jinja2~=3.1.4
packaging~=24.2
referencing~=0.35.1
MarkupSafe~=3.0.2
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
certifi~=2024.8.30
tzdata~=2024.2
tornado~=6.4.1
urllib3~=2.2.3
six~=1.16.0
idna~=3.10
tenacity~=9.0.0
watchdog~=5.0.3
click~=8.1.7
cachetools~=5.5.0
blinker~=1.9.0
jsonschema-specifications~=2024.10.1
pandas